use system timer for render timer, fix hide mouse issue
